scraper/requirements.txt
python-dotenv
pymongo
bs4
requests
pysocks